@@ -1,27 +1,28 @@
 use near_o11y::metrics::{exponential_buckets, try_create_histogram, Counter, Histogram};
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static PARTIAL_ENCODED_CHUNK_REQUEST_PROCESSING_TIME: Lazy<near_o11y::metrics::HistogramVec> =
-    Lazy::new(|| {
-        near_o11y::metrics::try_create_histogram_vec(
-            "near_partial_encoded_chunk_request_processing_time",
-            concat!(
-                "Time taken to prepare responses to partial encoded chuck ",
-                "requests.  The ‘method’ key describes how we tried to fulfil ",
-                "the request and ‘success’ describes whether we managed to ",
-                "create a response (‘ok’) or not (‘failed’).  Note that ",
-                "success does not mean that we managed to send the response ",
-                "over network; the count is taken before we attempt to send ",
-                "the data out."
-            ),
-            &["method", "success"],
-            Some(exponential_buckets(0.001, 2.0, 16).unwrap()),
-        )
-        .unwrap()
-    });
+pub static PARTIAL_ENCODED_CHUNK_REQUEST_PROCESSING_TIME: LazyLock<
+    near_o11y::metrics::HistogramVec,
+> = LazyLock::new(|| {
+    near_o11y::metrics::try_create_histogram_vec(
+        "near_partial_encoded_chunk_request_processing_time",
+        concat!(
+            "Time taken to prepare responses to partial encoded chuck ",
+            "requests.  The ‘method’ key describes how we tried to fulfil ",
+            "the request and ‘success’ describes whether we managed to ",
+            "create a response (‘ok’) or not (‘failed’).  Note that ",
+            "success does not mean that we managed to send the response ",
+            "over network; the count is taken before we attempt to send ",
+            "the data out."
+        ),
+        &["method", "success"],
+        Some(exponential_buckets(0.001, 2.0, 16).unwrap()),
+    )
+    .unwrap()
+});
 
-pub static DISTRIBUTE_ENCODED_CHUNK_TIME: Lazy<near_o11y::metrics::HistogramVec> =
-    Lazy::new(|| {
+pub static DISTRIBUTE_ENCODED_CHUNK_TIME: LazyLock<near_o11y::metrics::HistogramVec> =
+    LazyLock::new(|| {
         near_o11y::metrics::try_create_histogram_vec(
             "near_distribute_encoded_chunk_time",
             concat!(
@@ -34,7 +35,7 @@ pub static DISTRIBUTE_ENCODED_CHUNK_TIME: Lazy<near_o11y::metrics::HistogramVec>
         .unwrap()
     });
 
-pub(crate) static PARTIAL_ENCODED_CHUNK_RESPONSE_DELAY: Lazy<Histogram> = Lazy::new(|| {
+pub(crate) static PARTIAL_ENCODED_CHUNK_RESPONSE_DELAY: LazyLock<Histogram> = LazyLock::new(|| {
     try_create_histogram(
             "near_partial_encoded_chunk_response_delay",
             "Delay between when a partial encoded chunk response is sent from PeerActor and when it is received by ShardsManagerActor",
@@ -42,18 +43,20 @@ pub(crate) static PARTIAL_ENCODED_CHUNK_RESPONSE_DELAY: Lazy<Histogram> = Lazy::
         .unwrap()
 });
 
-pub static PARTIAL_ENCODED_CHUNK_FORWARD_CACHED_WITHOUT_HEADER: Lazy<Counter> = Lazy::new(|| {
-    near_o11y::metrics::try_create_counter(
+pub static PARTIAL_ENCODED_CHUNK_FORWARD_CACHED_WITHOUT_HEADER: LazyLock<Counter> = LazyLock::new(
+    || {
+        near_o11y::metrics::try_create_counter(
         "near_partial_encoded_chunk_forward_cached_without_header",
         concat!(
             "Number of times we received a valid partial encoded chunk forward without having the corresponding chunk header so we cached it"
         ),
     )
     .unwrap()
-});
+    },
+);
 
-pub static PARTIAL_ENCODED_CHUNK_FORWARD_CACHED_WITHOUT_PREV_BLOCK: Lazy<Counter> = Lazy::new(
-    || {
+pub static PARTIAL_ENCODED_CHUNK_FORWARD_CACHED_WITHOUT_PREV_BLOCK: LazyLock<Counter> =
+    LazyLock::new(|| {
         near_o11y::metrics::try_create_counter(
         "near_partial_encoded_chunk_forward_cached_without_prev_block",
         concat!(
@@ -61,5 +64,4 @@ pub static PARTIAL_ENCODED_CHUNK_FORWARD_CACHED_WITHOUT_PREV_BLOCK: Lazy<Counter
         ),
     )
     .unwrap()
-    },
-);
+    });
