@@ -1,7 +1,7 @@
 use near_o11y::metrics::{exponential_buckets, HistogramVec, IntCounter, IntCounterVec};
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static RPC_PROCESSING_TIME: Lazy<HistogramVec> = Lazy::new(|| {
+pub static RPC_PROCESSING_TIME: LazyLock<HistogramVec> = LazyLock::new(|| {
     near_o11y::metrics::try_create_histogram_vec(
         "near_rpc_processing_time",
         "Time taken to process rpc queries",
@@ -10,21 +10,21 @@ pub static RPC_PROCESSING_TIME: Lazy<HistogramVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub static RPC_TIMEOUT_TOTAL: Lazy<IntCounter> = Lazy::new(|| {
+pub static RPC_TIMEOUT_TOTAL: LazyLock<IntCounter> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_counter(
         "near_rpc_timeout_total",
         "Total count of rpc queries that ended on timeout",
     )
     .unwrap()
 });
-pub static PROMETHEUS_REQUEST_COUNT: Lazy<IntCounter> = Lazy::new(|| {
+pub static PROMETHEUS_REQUEST_COUNT: LazyLock<IntCounter> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_counter(
         "near_http_prometheus_requests_total",
         "Total count of Prometheus requests received",
     )
     .unwrap()
 });
-pub static HTTP_RPC_REQUEST_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
+pub static HTTP_RPC_REQUEST_COUNT: LazyLock<IntCounterVec> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_counter_vec(
         "near_rpc_total_count",
         "Total count of HTTP RPC requests received, by method",
@@ -32,14 +32,14 @@ pub static HTTP_RPC_REQUEST_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub static HTTP_STATUS_REQUEST_COUNT: Lazy<IntCounter> = Lazy::new(|| {
+pub static HTTP_STATUS_REQUEST_COUNT: LazyLock<IntCounter> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_counter(
         "near_http_status_requests_total",
         "Total count of HTTP Status requests received",
     )
     .unwrap()
 });
-pub static RPC_ERROR_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
+pub static RPC_ERROR_COUNT: LazyLock<IntCounterVec> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_counter_vec(
         "near_rpc_error_count",
         "Total count of errors by method and message",
@@ -47,7 +47,7 @@ pub static RPC_ERROR_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub static RPC_UNREACHABLE_ERROR_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
+pub static RPC_UNREACHABLE_ERROR_COUNT: LazyLock<IntCounterVec> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_counter_vec(
         "near_rpc_unreachable_errors_total",
         "Total count of Unreachable RPC errors returned, by target error enum",
