@@ -10,7 +10,7 @@ use near_o11y::metrics::{
     try_create_int_gauge, try_create_int_gauge_vec, Histogram, HistogramVec, IntCounter,
     IntCounterVec, IntGauge, IntGaugeVec, MetricVec, MetricVecBuilder,
 };
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
 /// Labels represents a schema of an IntGaugeVec metric.
 pub trait Labels: 'static {
@@ -107,18 +107,18 @@ impl<M: prometheus::core::Metric> std::ops::Deref for MetricGuard<M> {
 
 pub(crate) type IntGaugeGuard = MetricGuard<prometheus::IntGauge>;
 
-pub static PEER_CONNECTIONS: Lazy<Gauge<Connection>> =
-    Lazy::new(|| Gauge::new("near_peer_connections", "Number of connected peers").unwrap());
+pub static PEER_CONNECTIONS: LazyLock<Gauge<Connection>> =
+    LazyLock::new(|| Gauge::new("near_peer_connections", "Number of connected peers").unwrap());
 
-pub(crate) static PEER_CONNECTIONS_TOTAL: Lazy<IntGauge> = Lazy::new(|| {
+pub(crate) static PEER_CONNECTIONS_TOTAL: LazyLock<IntGauge> = LazyLock::new(|| {
     try_create_int_gauge("near_peer_connections_total", "Number of connected peers").unwrap()
 });
-pub(crate) static PEER_DATA_RECEIVED_BYTES: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static PEER_DATA_RECEIVED_BYTES: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter("near_peer_data_received_bytes", "Total data received from peers")
         .unwrap()
 });
 
-pub(crate) static PEER_MSG_SIZE_BYTES: Lazy<HistogramVec> = Lazy::new(|| {
+pub(crate) static PEER_MSG_SIZE_BYTES: LazyLock<HistogramVec> = LazyLock::new(|| {
     try_create_histogram_vec(
         "near_peer_msg_size_bytes",
         "Histogram of message sizes in bytes",
@@ -131,7 +131,7 @@ pub(crate) static PEER_MSG_SIZE_BYTES: Lazy<HistogramVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static PEER_MSG_READ_LATENCY: Lazy<Histogram> = Lazy::new(|| {
+pub(crate) static PEER_MSG_READ_LATENCY: LazyLock<Histogram> = LazyLock::new(|| {
     try_create_histogram_with_buckets(
         "near_peer_msg_read_latency",
         "Time that PeerActor spends on reading a message from a socket",
@@ -140,11 +140,11 @@ pub(crate) static PEER_MSG_READ_LATENCY: Lazy<Histogram> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static PEER_DATA_SENT_BYTES: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static PEER_DATA_SENT_BYTES: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter("near_peer_data_sent_bytes", "Total data sent to peers").unwrap()
 });
 
-pub(crate) static PEER_DATA_READ_BUFFER_SIZE: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub(crate) static PEER_DATA_READ_BUFFER_SIZE: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_peer_read_buffer_size",
         "Size of the message that this peer is currently sending to us",
@@ -152,7 +152,7 @@ pub(crate) static PEER_DATA_READ_BUFFER_SIZE: Lazy<IntGaugeVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub(crate) static PEER_DATA_WRITE_BUFFER_SIZE: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub(crate) static PEER_DATA_WRITE_BUFFER_SIZE: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_peer_write_buffer_size",
         "Size of the outgoing buffer for this peer",
@@ -160,23 +160,25 @@ pub(crate) static PEER_DATA_WRITE_BUFFER_SIZE: Lazy<IntGaugeVec> = Lazy::new(|| 
     )
     .unwrap()
 });
-pub(crate) static PEER_MESSAGE_RECEIVED_BY_TYPE_BYTES: Lazy<IntCounterVec> = Lazy::new(|| {
-    try_create_int_counter_vec(
-        "near_peer_message_received_by_type_bytes",
-        "Total data received from peers by message types",
-        &["type"],
-    )
-    .unwrap()
-});
-pub(crate) static PEER_MESSAGE_RECEIVED_BY_TYPE_TOTAL: Lazy<IntCounterVec> = Lazy::new(|| {
-    try_create_int_counter_vec(
-        "near_peer_message_received_by_type_total",
-        "Number of messages received from peers by message types",
-        &["type"],
-    )
-    .unwrap()
-});
-pub(crate) static PEER_MESSAGE_SENT_BY_TYPE_BYTES: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static PEER_MESSAGE_RECEIVED_BY_TYPE_BYTES: LazyLock<IntCounterVec> =
+    LazyLock::new(|| {
+        try_create_int_counter_vec(
+            "near_peer_message_received_by_type_bytes",
+            "Total data received from peers by message types",
+            &["type"],
+        )
+        .unwrap()
+    });
+pub(crate) static PEER_MESSAGE_RECEIVED_BY_TYPE_TOTAL: LazyLock<IntCounterVec> =
+    LazyLock::new(|| {
+        try_create_int_counter_vec(
+            "near_peer_message_received_by_type_total",
+            "Number of messages received from peers by message types",
+            &["type"],
+        )
+        .unwrap()
+    });
+pub(crate) static PEER_MESSAGE_SENT_BY_TYPE_BYTES: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_peer_message_sent_by_type_bytes",
         "Total data sent to peers by message types",
@@ -184,7 +186,7 @@ pub(crate) static PEER_MESSAGE_SENT_BY_TYPE_BYTES: Lazy<IntCounterVec> = Lazy::n
     )
     .unwrap()
 });
-pub(crate) static PEER_MESSAGE_SENT_BY_TYPE_TOTAL: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static PEER_MESSAGE_SENT_BY_TYPE_TOTAL: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_peer_message_sent_by_type_total",
         "Number of messages sent to peers by message types",
@@ -192,15 +194,16 @@ pub(crate) static PEER_MESSAGE_SENT_BY_TYPE_TOTAL: Lazy<IntCounterVec> = Lazy::n
     )
     .unwrap()
 });
-pub(crate) static PEER_MESSAGE_RATE_LIMITED_BY_TYPE_TOTAL: Lazy<IntCounterVec> = Lazy::new(|| {
-    try_create_int_counter_vec(
-        "near_peer_message_rate_limited_by_type_total",
-        "Number of messages dropped because rate limited by message types",
-        &["type"],
-    )
-    .unwrap()
-});
-pub(crate) static SYNC_ACCOUNTS_DATA: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static PEER_MESSAGE_RATE_LIMITED_BY_TYPE_TOTAL: LazyLock<IntCounterVec> =
+    LazyLock::new(|| {
+        try_create_int_counter_vec(
+            "near_peer_message_rate_limited_by_type_total",
+            "Number of messages dropped because rate limited by message types",
+            &["type"],
+        )
+        .unwrap()
+    });
+pub(crate) static SYNC_ACCOUNTS_DATA: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_sync_accounts_data",
         "Number of SyncAccountsData messages sent/received",
@@ -208,7 +211,7 @@ pub(crate) static SYNC_ACCOUNTS_DATA: Lazy<IntCounterVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub(crate) static SYNC_SNAPSHOT_HOSTS: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static SYNC_SNAPSHOT_HOSTS: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_sync_snapshot_hosts",
         "Number of SyncSnapshotHost messages sent/received",
@@ -217,7 +220,7 @@ pub(crate) static SYNC_SNAPSHOT_HOSTS: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static REQUEST_COUNT_BY_TYPE_TOTAL: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static REQUEST_COUNT_BY_TYPE_TOTAL: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_requests_count_by_type_total",
         "Number of network requests we send out, by message types",
@@ -227,31 +230,32 @@ pub(crate) static REQUEST_COUNT_BY_TYPE_TOTAL: Lazy<IntCounterVec> = Lazy::new(|
 });
 
 // Routing table metrics
-pub(crate) static ROUTING_TABLE_RECALCULATIONS: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static ROUTING_TABLE_RECALCULATIONS: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_routing_table_recalculations_total",
         "Number of times routing table have been recalculated from scratch",
     )
     .unwrap()
 });
-pub(crate) static ROUTING_TABLE_RECALCULATION_HISTOGRAM: Lazy<Histogram> = Lazy::new(|| {
-    try_create_histogram(
-        "near_routing_table_recalculation_seconds",
-        "Time spent recalculating routing table",
-    )
-    .unwrap()
-});
-pub(crate) static EDGE_UPDATES: Lazy<IntCounter> =
-    Lazy::new(|| try_create_int_counter("near_edge_updates", "Unique edge updates").unwrap());
-pub(crate) static EDGE_ACTIVE: Lazy<IntGauge> = Lazy::new(|| {
+pub(crate) static ROUTING_TABLE_RECALCULATION_HISTOGRAM: LazyLock<Histogram> =
+    LazyLock::new(|| {
+        try_create_histogram(
+            "near_routing_table_recalculation_seconds",
+            "Time spent recalculating routing table",
+        )
+        .unwrap()
+    });
+pub(crate) static EDGE_UPDATES: LazyLock<IntCounter> =
+    LazyLock::new(|| try_create_int_counter("near_edge_updates", "Unique edge updates").unwrap());
+pub(crate) static EDGE_ACTIVE: LazyLock<IntGauge> = LazyLock::new(|| {
     try_create_int_gauge("near_edge_active", "Total edges active between peers").unwrap()
 });
-pub(crate) static EDGE_TOTAL: Lazy<IntGauge> = Lazy::new(|| {
+pub(crate) static EDGE_TOTAL: LazyLock<IntGauge> = LazyLock::new(|| {
     try_create_int_gauge("near_edge_total", "Total edges between peers (including removed ones).")
         .unwrap()
 });
 
-pub(crate) static EDGE_TOMBSTONE_SENDING_SKIPPED: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static EDGE_TOMBSTONE_SENDING_SKIPPED: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_edge_tombstone_sending_skip",
         "Number of times that we didn't send tombstones.",
@@ -259,7 +263,7 @@ pub(crate) static EDGE_TOMBSTONE_SENDING_SKIPPED: Lazy<IntCounter> = Lazy::new(|
     .unwrap()
 });
 
-pub(crate) static EDGE_TOMBSTONE_RECEIVING_SKIPPED: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static EDGE_TOMBSTONE_RECEIVING_SKIPPED: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_edge_tombstone_receiving_skip",
         "Number of times that we pruned tombstones upon receiving.",
@@ -267,14 +271,14 @@ pub(crate) static EDGE_TOMBSTONE_RECEIVING_SKIPPED: Lazy<IntCounter> = Lazy::new
     .unwrap()
 });
 
-pub(crate) static PEER_UNRELIABLE: Lazy<IntGauge> = Lazy::new(|| {
+pub(crate) static PEER_UNRELIABLE: LazyLock<IntGauge> = LazyLock::new(|| {
     try_create_int_gauge(
         "near_peer_unreliable",
         "Total peers that are behind and will not be used to route messages",
     )
     .unwrap()
 });
-pub(crate) static PEER_MANAGER_TRIGGER_TIME: Lazy<HistogramVec> = Lazy::new(|| {
+pub(crate) static PEER_MANAGER_TRIGGER_TIME: LazyLock<HistogramVec> = LazyLock::new(|| {
     try_create_histogram_vec(
         "near_peer_manager_trigger_time",
         "Time that PeerManagerActor spends on different types of triggers",
@@ -283,7 +287,7 @@ pub(crate) static PEER_MANAGER_TRIGGER_TIME: Lazy<HistogramVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub(crate) static PEER_MANAGER_MESSAGES_TIME: Lazy<HistogramVec> = Lazy::new(|| {
+pub(crate) static PEER_MANAGER_MESSAGES_TIME: LazyLock<HistogramVec> = LazyLock::new(|| {
     try_create_histogram_vec(
         "near_peer_manager_messages_time",
         "Time that PeerManagerActor spends on handling different types of messages",
@@ -292,7 +296,7 @@ pub(crate) static PEER_MANAGER_MESSAGES_TIME: Lazy<HistogramVec> = Lazy::new(|| 
     )
     .unwrap()
 });
-pub(crate) static ROUTED_MESSAGE_DROPPED: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static ROUTED_MESSAGE_DROPPED: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_routed_message_dropped",
         "Number of messages dropped due to TTL=0, by routed message type",
@@ -301,14 +305,14 @@ pub(crate) static ROUTED_MESSAGE_DROPPED: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static PEER_REACHABLE: Lazy<IntGauge> = Lazy::new(|| {
+pub(crate) static PEER_REACHABLE: LazyLock<IntGauge> = LazyLock::new(|| {
     try_create_int_gauge(
         "near_peer_reachable",
         "Total peers such that there is a path potentially through other peers",
     )
     .unwrap()
 });
-static DROPPED_MESSAGE_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
+static DROPPED_MESSAGE_COUNT: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_dropped_message_by_type_and_reason_count",
         "Total count of messages which were dropped by type of message and \
@@ -317,7 +321,7 @@ static DROPPED_MESSAGE_COUNT: Lazy<IntCounterVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-pub(crate) static PARTIAL_ENCODED_CHUNK_REQUEST_DELAY: Lazy<Histogram> = Lazy::new(|| {
+pub(crate) static PARTIAL_ENCODED_CHUNK_REQUEST_DELAY: LazyLock<Histogram> = LazyLock::new(|| {
     try_create_histogram(
         "near_partial_encoded_chunk_request_delay",
         "Delay between when a partial encoded chunk request is sent from ClientActor and when it is received by PeerManagerActor",
@@ -325,11 +329,11 @@ pub(crate) static PARTIAL_ENCODED_CHUNK_REQUEST_DELAY: Lazy<Histogram> = Lazy::n
         .unwrap()
 });
 
-pub(crate) static BROADCAST_MESSAGES: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static BROADCAST_MESSAGES: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec("near_broadcast_msg", "Broadcasted messages", &["type"]).unwrap()
 });
 
-static NETWORK_ROUTED_MSG_LATENCY: Lazy<HistogramVec> = Lazy::new(|| {
+static NETWORK_ROUTED_MSG_LATENCY: LazyLock<HistogramVec> = LazyLock::new(|| {
     try_create_histogram_vec(
         "near_network_routed_msg_latency",
         "Latency of network messages, assuming clocks are perfectly synchronized. 'tier' indicates what is the tier of the connection on which the message arrived (TIER1 is expected to be faster than TIER2) and 'fastest' indicates whether this was the first copy of the message to arrive.",
@@ -338,7 +342,7 @@ static NETWORK_ROUTED_MSG_LATENCY: Lazy<HistogramVec> = Lazy::new(|| {
     )
     .unwrap()
 });
-static NETWORK_ROUTED_MSG_NUM_HOPS: Lazy<IntCounterVec> = Lazy::new(|| {
+static NETWORK_ROUTED_MSG_NUM_HOPS: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_network_routed_msg_hops",
         "Number of peers the routed message travelled through",
@@ -347,7 +351,7 @@ static NETWORK_ROUTED_MSG_NUM_HOPS: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static CONNECTED_TO_MYSELF: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static CONNECTED_TO_MYSELF: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_connected_to_myself",
         "This node connected to itself, this shouldn't happen",
@@ -355,7 +359,7 @@ pub(crate) static CONNECTED_TO_MYSELF: Lazy<IntCounter> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static ALREADY_CONNECTED_ACCOUNT: Lazy<IntCounter> = Lazy::new(|| {
+pub(crate) static ALREADY_CONNECTED_ACCOUNT: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_already_connected_account",
         "A second peer with the same validator key is trying to connect to our node. This means that the validator peer has invalid setup."
@@ -363,7 +367,7 @@ pub(crate) static ALREADY_CONNECTED_ACCOUNT: Lazy<IntCounter> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static ACCOUNT_TO_PEER_LOOKUPS: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static ACCOUNT_TO_PEER_LOOKUPS: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_account_to_peer_lookups",
         "number of lookups of peer_id by account_id (for routed messages)",
@@ -376,7 +380,7 @@ pub(crate) static ACCOUNT_TO_PEER_LOOKUPS: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static NETWORK_ROUTED_MSG_DISTANCES: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static NETWORK_ROUTED_MSG_DISTANCES: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_network_routed_msg_distances",
         "compares routing distance by protocol (V1 vs V2)",
