@@ -1,7 +1,7 @@
 use near_o11y::metrics::IntGaugeVec;
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static TRANSACTION_POOL_COUNT: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static TRANSACTION_POOL_COUNT: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_gauge_vec(
         "near_transaction_pool_entries",
         "Total number of transactions currently tracked by the node in a given shard pool",
@@ -10,7 +10,7 @@ pub static TRANSACTION_POOL_COUNT: Lazy<IntGaugeVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub static TRANSACTION_POOL_SIZE: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static TRANSACTION_POOL_SIZE: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     near_o11y::metrics::try_create_int_gauge_vec(
         "near_transaction_pool_size",
         "Total size in bytes of transactions currently tracked by the node in a given shard pool",
