@@ -1,7 +1,7 @@
 use near_o11y::metrics::{try_create_int_gauge_vec, IntGaugeVec};
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static CONFIG_MUTABLE_FIELD: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static CONFIG_MUTABLE_FIELD: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_config_mutable_field",
         "Timestamp and value of a mutable config field",
