@@ -1,7 +1,7 @@
 use near_o11y::metrics::{try_create_int_gauge_vec, IntGaugeVec};
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static MEM_TRIE_ARENA_ACTIVE_ALLOCS_BYTES: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static MEM_TRIE_ARENA_ACTIVE_ALLOCS_BYTES: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_mem_trie_arena_active_allocs_bytes",
         "Total size of active allocations on the in-memory trie arena",
@@ -10,7 +10,7 @@ pub static MEM_TRIE_ARENA_ACTIVE_ALLOCS_BYTES: Lazy<IntGaugeVec> = Lazy::new(|| 
     .unwrap()
 });
 
-pub static MEM_TRIE_ARENA_ACTIVE_ALLOCS_COUNT: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static MEM_TRIE_ARENA_ACTIVE_ALLOCS_COUNT: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_mem_trie_arena_active_allocs_count",
         "Total number of active allocations on the in-memory trie arena",
@@ -19,7 +19,7 @@ pub static MEM_TRIE_ARENA_ACTIVE_ALLOCS_COUNT: Lazy<IntGaugeVec> = Lazy::new(|| 
     .unwrap()
 });
 
-pub static MEM_TRIE_ARENA_MEMORY_USAGE_BYTES: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static MEM_TRIE_ARENA_MEMORY_USAGE_BYTES: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_mem_trie_arena_memory_usage_bytes",
         "Memory usage of the in-memory trie arena",
