@@ -2,9 +2,9 @@ use near_o11y::metrics::{
     try_create_int_counter, try_create_int_counter_vec, try_create_int_gauge_vec, IntCounter,
     IntCounterVec, IntGaugeVec,
 };
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static MEM_TRIE_NUM_ROOTS: Lazy<IntGaugeVec> = Lazy::new(|| {
+pub static MEM_TRIE_NUM_ROOTS: LazyLock<IntGaugeVec> = LazyLock::new(|| {
     try_create_int_gauge_vec(
         "near_mem_trie_num_roots",
         "Number of trie roots currently active in the in-memory trie",
@@ -13,7 +13,7 @@ pub static MEM_TRIE_NUM_ROOTS: Lazy<IntGaugeVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub static MEM_TRIE_NUM_NODES_CREATED_FROM_UPDATES: Lazy<IntCounterVec> = Lazy::new(|| {
+pub static MEM_TRIE_NUM_NODES_CREATED_FROM_UPDATES: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_mem_trie_num_nodes_created_from_updates",
         "Number of trie nodes created by applying updates",
@@ -22,7 +22,7 @@ pub static MEM_TRIE_NUM_NODES_CREATED_FROM_UPDATES: Lazy<IntCounterVec> = Lazy::
     .unwrap()
 });
 
-pub static MEM_TRIE_NUM_LOOKUPS: Lazy<IntCounter> = Lazy::new(|| {
+pub static MEM_TRIE_NUM_LOOKUPS: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_mem_trie_num_lookups",
         "Number of in-memory trie lookups (number of keys looked up)",
