@@ -1,9 +1,9 @@
 use near_o11y::metrics::{
     try_create_int_counter, try_create_int_counter_vec, IntCounter, IntCounterVec,
 };
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub static TRANSACTIONS_SENT: Lazy<IntCounterVec> = Lazy::new(|| {
+pub static TRANSACTIONS_SENT: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "near_mirror_transactions_sent",
         "Total number of transactions sent",
@@ -12,7 +12,7 @@ pub static TRANSACTIONS_SENT: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub static TRANSACTIONS_INCLUDED: Lazy<IntCounter> = Lazy::new(|| {
+pub static TRANSACTIONS_INCLUDED: LazyLock<IntCounter> = LazyLock::new(|| {
     try_create_int_counter(
         "near_mirror_transactions_included",
         "Total number of transactions sent that made it on-chain",
