@@ -2,9 +2,9 @@ use near_o11y::metrics::{
     exponential_buckets, try_create_histogram_vec, try_create_int_counter_vec, HistogramVec,
     IntCounterVec,
 };
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
-pub(crate) static PONG_RECEIVED: Lazy<HistogramVec> = Lazy::new(|| {
+pub(crate) static PONG_RECEIVED: LazyLock<HistogramVec> = LazyLock::new(|| {
     try_create_histogram_vec(
         "ping_pong_received",
         "Round-trip time of ping-pong",
@@ -14,7 +14,7 @@ pub(crate) static PONG_RECEIVED: Lazy<HistogramVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static PONG_TIMEOUTS: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static PONG_TIMEOUTS: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "ping_pong_timeout",
         "Number of pongs that were not received",
@@ -23,7 +23,7 @@ pub(crate) static PONG_TIMEOUTS: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-pub(crate) static PING_SENT: Lazy<IntCounterVec> = Lazy::new(|| {
+pub(crate) static PING_SENT: LazyLock<IntCounterVec> = LazyLock::new(|| {
     try_create_int_counter_vec(
         "ping_ping_sent",
         "Number of pings sent",
